Rename create{ -> UserSubscription}TableIfNotExists.js
