+++ prelaunch/createUserSubscriptionTableIfNotExists.js
